refactored settings menu extension
